wip
Improve the Route53 portion so that it is an "ALIAS" to the ELB. This
effort began to be unwieldy with method parameters growing in length so
I've created a somewhat less offensive `AWSLoadBalancer` structure.

Maybe at some point we ought to start dealing in native AWS SDK objects,
but we'll get there maybe.

Signed-off-by: Lisa Seelye <lseelye@redhat.com>