OSD-3190: Add unit tests for utils
This adds the "testutils" helper package which is in use for the added
tests. Because there was going to be duplication in the apischeme
controller tests, those were consolidated and updated to reflect the new
helper.

Notably, utils/conditions.go is not tested (yet), and publishingstrategy
controller tests aren't unified with the helper at this time.

Signed-off-by: Lisa Seelye <lisa@users.noreply.github.com>