OSD-3564: Remove extra log statements & Spec is mandatory
This change is wider scoped than only APIScheme because it also makes
Spec mandatory for PublishingStrategy CRDs. We always want these to be
mandatory since we need them present to do anything useful in their
respective controllers.

In order to keep tabs on some of what is going on in the APIScheme
controller, there is a new variable to control when to print it's
activities. It will log activity every 1000 reconcilations (every ~1000
minutes), just so we have some other kind of heartbeat.

Signed-off-by: Lisa Seelye <lisa@users.noreply.github.com>