Make enableManagementAPIServerIngress variable
As part of the solution for the referenced Jira, we want v3 hives to
*not* apply the APIScheme until private-cluster-rhapi-apischeme-updater
has gotten a chance to run. So we're making that a variable, which we'll
set to `false` for v3 hives and `true` for v4 hives in the appsre
pipeline. The private-cluster-rhapi-apischeme-updater will then toggle
it to `true` as it updates the allowedCIDRBlocks.

Jira: [OHSS-909](https://issues.redhat.com/browse/OHSS-909)