sshd: Explicitly make host keys group readable
The sshd process runs as a non-root user within the pod, whereas the
mounted host keys are owned by root.  We therefore depend on the key
files being group-readable.  Set this read bit explicitly instead of
depending on the security context constraint (SCC) to set it.

Normally group-readable private keys are forbidden by OpenSSH, but
it turns out the server does NOT apply its strict permission checks
when a private key is owned by a different user than itself.