OSD-3566: Introduce cloudclient interface
This change will introduce several new packages:

* pkg/cloudclient
* pkg/cloudclient/aws
* pkg/errors

The first defines a common interface each cloud client will need to
satisfy to perform functions related to what cloud-ingress-operator does
with respect to cloud service providers (such as AWS). The second
package is the in-progress AWS implementation. The third package is a
set of bespoke errors used by the operator (mainly the cloudclient
implementations).